carriage return management
When downloading a file via maven, a count down from 0% to 100% is done. This is done by using the \r character (carriage return). But it seems that ruby automatically convert that character to \n (line feed) when reading the lines via the each function.
Hence, instead of reading line by line, the file is read byte by byte.